requirements.txt
python-dotenv
numpy==1.26.4
openai==1.57.0
pandas==1.5.3
pandasai==2.4.0
seaborn==0.13.2
streamlit==1.40.2
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
nltk==3.9.1
textblob==0.18.0
wordcloud==1.9.4
ipykernel==6.29.5
nbformat==5.10.4
ipwidgets==8.1.5
